add to cart with quantity working

+++ src/app/components/product/product-list/product-list.component.ts
@@ -5,7 +5,10 @@ import { Category } from 'src/app/models/product/Category';
 import { ActivatedRoute } from '@angular/router';
 import { CategoryService } from 'src/app/services/product/category.service';
 import { CartService } from 'src/app/services/customer/cart.service';
-
+import sweetalert2 from 'sweetalert2'
+ 
+// CommonJS
+const Swal = require('sweetalert2');
 @Component({
   selector: 'app-product-list',
   templateUrl: './product-list.component.html',
@@ -111,14 +114,47 @@ export class ProductListComponent implements OnInit {
   // }
   
   addProductToCart(productId:number){
-    if (localStorage.getItem("CartToken")==null){
-      this.cartService.createAnonymCart().subscribe(res=>{
-        localStorage.setItem("CartToken",res);
-        this.cartService.addProductToCart(productId).subscribe()});
-    }
-    else {
-      this.cartService.addProductToCart(productId).subscribe();
-    }
+
+    Swal.fire({
+      title: 'Are you sure?',
+      text: 'Do you confirm adding this product to your cart !',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Confirm',
+      cancelButtonText: 'Cancel'
+    }).then((result) => {
+      if (result.value) {
+        //if agree
+        if (localStorage.getItem("CartToken")==null){
+          this.cartService.createAnonymCart().subscribe(res=>{
+            localStorage.setItem("CartToken",res);
+            this.cartService.addProductToCart(productId).subscribe()});
+        }
+        else {
+          this.cartService.addProductToCart(productId).subscribe();
+        }
+        Swal.fire(
+          'Success !',
+          'Product successfuly added to cart',
+          'success'
+        )
+      // For more information about handling dismissals please visit
+      // https://sweetalert2.github.io/#handling-dismissals
+      } else if (result.dismiss === Swal.DismissReason.cancel) {
+        Swal.fire(
+          'Cancelled',
+          'Request canceled :)',
+          'Canceled'
+        )
+      }
+    })
+
+    
+
+
+
+
+   
 
     
   }
